rename new_needs_selector
