+++ __tests__/components/needs_selector_test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { mount } from "enzyme";
-import NeedsSelector from "../../components/new_needs_selector";
+import NeedsSelector from "../../components/needs_selector";
 import needsFixture from "../fixtures/needs";
 
 describe("NeedsSelector", () => {
